--- src/utils/constants/Info.ts
@@ -1,4 +1,0 @@
-export const INFO = {
-	OWNER: 'facebook',
-	REPO: 'react',
-};
