+++ src/utils/constants/constants.ts
@@ -8,3 +8,8 @@ export const ROUTES = {
 	WANTED_IMG:
 		'https://image.wanted.co.kr/optimize?src=https%3A%2F%2Fstatic.wanted.co.kr%2Fimages%2Fuserweb%2Flogo_wanted_black.png&w=110&q=100',
 };
+
+export const INFO = {
+	OWNER: 'facebook',
+	REPO: 'react',
+};
